requirements.txt
python-telegram-bot==20.0
solana==0.21.0
